+++ src/Pim/Bundle/CatalogBundle/Tests/Unit/MassEditAction/EditCommonAttributesTest.php
@@ -62,27 +62,27 @@ class EditCommonAttributesTest extends \PHPUnit_Framework_TestCase
             ->will($this->returnValue(array($sku, $name, $color, $description, $price)));
 
         $foo->expects($this->any())
-            ->method('getValue')
+            ->method('hasAttribute')
             ->will(
                 $this->returnValueMap(
                     array(
-                        array('name',         null,  null,  true),
-                        array('color',        null,  null,  true),
-                        array('description',  null,  null,  true),
-                        array('price',        null,  null,  true),
+                        array($name,        true),
+                        array($color,       true),
+                        array($description, true),
+                        array($price,       true),
                     )
                 )
             );
 
         $bar->expects($this->any())
-            ->method('getValue')
+            ->method('hasAttribute')
             ->will(
                 $this->returnValueMap(
                     array(
-                        array('name',         null,  null,  true),
-                        array('color',        null,  null,  false),
-                        array('description',  null,  null,  true),
-                        array('price',        null,  null,  true),
+                        array($name,        true),
+                        array($color,       false),
+                        array($description, true),
+                        array($price,       true),
                     )
                 )
             );