+++ src/Pim/Bundle/ImportExportBundle/Transformer/ORMTransformer.php
@@ -2,6 +2,8 @@
 
 namespace Pim\Bundle\ImportExportBundle\Transformer;
 
+use Pim\Bundle\ImportExportBundle\Exception\MissingIdentifierException;
+
 /**
  * Transforms an array in an entity
  *
@@ -31,7 +33,7 @@ class ORMTransformer extends AbstractORMTransformer
     protected function getEntity($class, array $data)
     {
         if (!isset($data['code'])) {
-            throw new MissingIdentifierException;
+            throw new MissingIdentifierException();
         }
         $object = $this->doctrine->getRepository($class)->findOneBy(array('code' => $data['code']));
         if (!$object) {